Add CNN based AE
