+++ energy_fault_detector/autoencoders/__init__.py
@@ -2,5 +2,5 @@
 
 from .multilayer_autoencoder import MultilayerAutoencoder
 from .conditional_autoencoder import ConditionalAE
-from .lstm_seq2seq_autoencoder import LSTMSeqAutoencoder
 from .lstm_seq2one_autoencoder import LSTMSeq2OneAutoencoder
+from .cnn_seq2one_autoencoder import CNNSeq2OneAutoencoder
